Refactor most builder logic to TestRailAPIAdapter
This has a few benefits:
* Simplifies the dependency hierarchy
* Hides the TestRailRequestStats exclusively behind the TestRailAPIAdapter
* Allows the TestRailAPIObjectBuilder to focus on building more complex object structures
* TestRailAPIAdapter performs the deletion, allowing the DeletionHandlers to better obscure the implementation
